+++ AddOns/Platynator/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Platynator
 
-## [344](https://github.com/TheMouseNest/Platynator/tree/344) (2026-03-13)
-[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/343...344) [Previous Releases](https://github.com/TheMouseNest/Platynator/releases)
+## [345](https://github.com/TheMouseNest/Platynator/tree/345) (2026-03-14)
+[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/344...345) [Previous Releases](https://github.com/TheMouseNest/Platynator/releases)
 
-- Add version slash command  
+- Fix typo that caused cannot interrupt marker to sometimes hide  
+- Handle cannot interrupt marker not showing if a spell changes to uninterruptible  
+- Use more reliable solution for purging realm names from Blizzard friendly plates  
+- Retail: Remove realm names from friendly plates in instances